+++ src/HareDu.Diagnostics.Tests/Fakes/FakeChannelSnapshot2.cs
@@ -18,17 +18,17 @@ namespace HareDu.Diagnostics.Tests.Fakes
     public class FakeChannelSnapshot2 :
         ChannelSnapshot
     {
-        public FakeChannelSnapshot2(long prefetchCount)
+        public FakeChannelSnapshot2(uint prefetchCount)
         {
             PrefetchCount = prefetchCount;
         }
 
-        public long PrefetchCount { get; }
-        public long UncommittedAcknowledgements { get; }
-        public long UncommittedMessages { get; }
-        public long UnconfirmedMessages { get; }
-        public long UnacknowledgedMessages { get; }
-        public long Consumers { get; }
+        public uint PrefetchCount { get; }
+        public ulong UncommittedAcknowledgements { get; }
+        public ulong UncommittedMessages { get; }
+        public ulong UnconfirmedMessages { get; }
+        public ulong UnacknowledgedMessages { get; }
+        public ulong Consumers { get; }
         public string Name { get; }
         public string Node { get; }
     }
